rev: Rename `config` module to `group`
- Clarify module role ...
