+++ config-it/src/beacon/group.rs
