+++ config-it/src/beacon/mod.rs
@@ -1,4 +1,4 @@
-pub mod config;
 pub mod entity;
+pub mod group;
 pub mod noti;
 pub mod storage;
